--- CSDN/src/META-INF/MANIFEST.txt
@@ -1,3 +1,0 @@
-Manifest-Version: 1.0
-Main-Class: Main
-
