--- README.md
@@ -1,2 +1,0 @@
-# CSDN-always-resquet
-用Java网络编程的Api写了一个刷取CSDN访问量的脚本 目前放在服务器里面跑 JDK1.8 使用到的是net包,io包
