+++ src/test/java/org/example/csdnalwaysrequestv1/CsdnAlwaysRequestV1ApplicationTests.java
@@ -1,0 +1,13 @@
+package org.example.csdnalwaysrequestv1;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CsdnAlwaysRequestV1ApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
